refactored the cli module into smaller modules
