+++ cffconvert/__init__.py
@@ -1,8 +1,6 @@
 from cffconvert.citation import Citation
-from cffconvert.cli import cli
 
 
 __all__ = [
-    "Citation",
-    "cli"
+    "Citation"
 ]
