+++ cffconvert/cli/read_from_file.py
@@ -1,0 +1,3 @@
+def read_from_file(infile):
+    with open(infile, "r", encoding="utf8") as f:
+        return f.read()
