+++ setup.cfg
@@ -59,7 +59,7 @@ install_requires =
 
 [options.entry_points]
 console_scripts =
-    cffconvert = cffconvert:cli
+    cffconvert = cffconvert.cli.cli:cli
 
 [options.extras_require]
 dev =
